Develop to master.

+++ Client/Settings/CameraVideoSourceSettings.h
@@ -2,7 +2,7 @@
 #include <nlohmann/json.hpp>
 struct CameraVideoSourceSettings
 {
-	int index{ 0 };
+	size_t index{ 0 };
 	bool gray{ false };
 };
 NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE(CameraVideoSourceSettings, index, gray)
